+++ compiler/qsc/src/interpret/circuit_tests.rs
@@ -603,26 +603,6 @@ fn lambda() {
         namespace Test {
             @EntryPoint()
             operation Main() : Result[] { [] }
-
-            operation SWAP (q1 : Qubit, q2 : Qubit) : Unit
-                is Adj + Ctl {
-
-                body (...) {
-                    CNOT(q1, q2);
-                    CNOT(q2, q1);
-                    CNOT(q1, q2);
-                }
-
-                adjoint (...) {
-                    SWAP(q1, q2);
-                }
-
-                controlled (cs, ...) {
-                    CNOT(q1, q2);
-                    Controlled CNOT(cs, (q2, q1));
-                    CNOT(q1, q2);
-                }
-            }
         }",
         Profile::Unrestricted,
     );
